requirements.txt
aiohttp==3.10.10
APScheduler==3.10.4
matplotlib==3.9.2
mplfinance==0.12.10b0
numpy==2.1.2
pandas==2.2.3
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
TA-Lib==0.4.32
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
cryptography==46.0.2
python-dateutil==2.9.0.post0
six==1.17.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0